# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FraudTransaction {
  id: number;
  txnRef: string;
  amount: number;
  category: string;
  hour: number;
  distance: number;
  frequency: number;
  deviceType: string;
  country: string;
  isFirstTransaction: boolean;
  score: number;
  status: string;
  velocityRisk: number;
  geographicRisk: number;
  behavioralRisk: number;
  deviceRisk: number;
  amlStatus: string;
  complianceTier: string;
  createdAt: string;
}

export interface TransactionInput {
  txnRef: string;
  amount: number;
  category: string;
  hour: number;
  distance: number;
  frequency: number;
  deviceType: string;
  country: string;
  isFirstTransaction: boolean;
  score: number;
  status: string;
  velocityRisk: number;
  geographicRisk: number;
  behavioralRisk: number;
  deviceRisk: number;
  amlStatus: string;
  complianceTier: string;
}

export interface TransactionListResponse {
  transactions: FraudTransaction[];
  total: number;
}

export interface CategoryStat {
  category: string;
  count: number;
  avgScore: number;
}

export interface HourlyTrendPoint {
  hour: string;
  count: number;
}

export interface TransactionStats {
  totalAnalyzed: number;
  totalBlocked: number;
  totalReview: number;
  totalApproved: number;
  avgScore: number;
  totalAmountProtected: number;
  fraudByCategory: CategoryStat[];
  hourlyTrend: HourlyTrendPoint[];
}

export interface DeleteResult {
  success: boolean;
}

export interface BulkTransactionInput {
  transactions: TransactionInput[];
}

export interface BulkCreateResult {
  saved: number;
  failed: number;
  total: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListTransactionsParams = {
status?: string;
category?: string;
search?: string;
minScore?: number;
maxScore?: number;
limit?: number;
offset?: number;
};