# lib/api-zod/src/generated/types/bulkTransactionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionInput } from './transactionInput';

export interface BulkTransactionInput {
  transactions: TransactionInput[];
}